Refactor to use new memory and base modules - maintain backward compatibility
